VLM_visual_research/requirements.txt
torch==2.4.0
accelerate==0.29.1
datasets==2.20.0
flash-attn==2.6.3
huggingface-hub==0.22.2
numpy==1.26.4
peft==0.11.1
pillow==10.4.0
torchvision==0.19.0
tqdm==4.66.4
transformers==4.46.3
wandb==0.17.7
opencv-python